Latest small changes

+++ src/arguments.h
@@ -16,7 +16,7 @@ public:
 
         modulePath = argv[0];
 
-        args.reserve(argc-1);
+        args.reserve((size_t)(argc-1));
         for (int i = 1; i < argc; i++)
         {
             std::string arg = argv[i];
